@@ -38,46 +38,46 @@
     "sensio/framework-extra-bundle": "^6.2",
     "sentry/sentry-symfony": "^4.2",
     "stof/doctrine-extensions-bundle": "^1.5",
-    "symfony/asset": "5.4.*",
-    "symfony/browser-kit": "5.4.*",
-    "symfony/console": "5.4.*",
-    "symfony/debug-bundle": "5.4.*",
-    "symfony/dom-crawler": "5.4.*",
-    "symfony/dotenv": "5.4.*",
+    "symfony/asset": "6.1.*",
+    "symfony/browser-kit": "6.1.*",
+    "symfony/console": "6.1.*",
+    "symfony/debug-bundle": "6.1.*",
+    "symfony/dom-crawler": "6.1.*",
+    "symfony/dotenv": "6.1.*",
     "symfony/flex": "^1.17",
-    "symfony/form": "5.4.*",
-    "symfony/framework-bundle": "5.4.*",
-    "symfony/http-client": "5.4.*",
-    "symfony/mailer": "5.4.*",
+    "symfony/form": "6.1.*",
+    "symfony/framework-bundle": "6.1.*",
+    "symfony/http-client": "6.1.*",
+    "symfony/mailer": "6.1.*",
     "symfony/maker-bundle": "^1.47",
-    "symfony/messenger": "5.4.*",
-    "symfony/mime": "5.4.*",
+    "symfony/messenger": "6.1.*",
+    "symfony/mime": "6.1.*",
     "symfony/monolog-bundle": "^3.0",
-    "symfony/property-access": "5.4.*",
-    "symfony/property-info": "5.4.*",
-    "symfony/runtime": "5.4.*",
-    "symfony/security-bundle": "5.4.*",
-    "symfony/serializer": "5.4.*",
-    "symfony/stopwatch": "5.4.*",
-    "symfony/twig-bundle": "5.4.*",
-    "symfony/validator": "5.4.*",
-    "symfony/var-dumper": "5.4.*",
-    "symfony/web-profiler-bundle": "5.4.*",
+    "symfony/property-access": "6.1.*",
+    "symfony/property-info": "6.1.*",
+    "symfony/runtime": "6.1.*",
+    "symfony/security-bundle": "6.1.*",
+    "symfony/serializer": "6.1.*",
+    "symfony/stopwatch": "6.1.*",
+    "symfony/twig-bundle": "6.1.*",
+    "symfony/validator": "6.1.*",
+    "symfony/var-dumper": "6.1.*",
+    "symfony/web-profiler-bundle": "6.1.*",
     "symfony/webpack-encore-bundle": "^1.8",
-    "symfony/yaml": "5.4.*",
+    "symfony/yaml": "6.1.*",
     "twig/extra-bundle": "^2.12|^3.0",
     "twig/twig": "^2.12|^3.0"
   },
   "require-dev": {
     "dama/doctrine-test-bundle": "^v6.6.0",
     "doctrine/doctrine-fixtures-bundle": "^3.3",
-    "liip/test-fixtures-bundle": "2.0.1",
     "phpstan/phpstan": "^1.1",
     "phpunit/phpunit": "^9.5",
     "rector/rector": "^0.12.5",
-    "symfony/css-selector": "5.3.*",
+    "symfony/css-selector": "6.1.*",
     "symfony/phpunit-bridge": "^6.1",
-    "symplify/easy-coding-standard": "^9"
+    "symplify/easy-coding-standard": "^9",
+    "zenstruck/foundry": "^1.22"
   },
   "config": {
     "optimize-autoloader": true,
@@ -127,7 +127,7 @@
   "extra": {
     "symfony": {
       "allow-contrib": false,
-      "require": "5.4.*"
+      "require": "6.1.*"
     }
   }
 }
